requirements.txt
certifi==2024.2.2
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
tenacity==9.0.0
urllib3==2.2.1

